Add section to explain first steps/how this works.
This should at least contain a link to an explanation how these samples
work in general, e.g. what the dev/config/tools folders are, how to
start the samples (should be similar) and general conventions in
pedestal. In particular, I think that it should be pointed out that you
can play with almost all parts in the repl and how a pedestal app is put
together.